LayoutTests/http/wpt/cross-origin-options/navigation-from-opener-via-open-target.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Tests that 'Cross-Origin-Options: deny / allow-postmessage' prevents a cross-origin opener from navigating us</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/utils.js"></script>
</head>
<body>
<script>

promise_test(t => {
    return withPopup("serve-cross-origin-options-header.py?value=deny", true /* isCrossOrigin */, "foo1").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = (msg) => {
                assert_not_equals(msg.source, result.window, "Existing window should not navigate");
            }

            let destinationURL = get_host_info().HTTP_ORIGIN + "/WebKit/cross-origin-options/resources/destination.html";
            w = open(destinationURL, "foo1");
            // If a window with the given name is found but cannot be navigated, a new one is created, as if we could
            // not find the given window.
            assert_not_equals(w, result.window, "open() should a new window");

            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: deny' prevents navigation from opener via open() target");

promise_test(t => {
    return withPopup("serve-cross-origin-options-header.py?value=allow-postmessage", true /* isCrossOrigin */, "foo2").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = (msg) => {
                assert_not_equals(msg.source, result.window, "Existing window should not navigate");
            }

            let destinationURL = get_host_info().HTTP_ORIGIN + "/WebKit/cross-origin-options/resources/destination.html";
            w = open(destinationURL, "foo2");
            // If a window with the given name is found but cannot be navigated, a new one is created, as if we could
            // not find the given window.
            assert_not_equals(w, result.window, "open() should a new window");

            t.step_timeout(() => {
               window.onmessage = null;
               resolve();
            }, 200);
        });
    });
}, "'Cross-Origin-Options: allow-postmessage' prevents navigation from opener via open() target");

promise_test(t => {
    return withPopup("serve-cross-origin-options-header.py?value=allow", true /* isCrossOrigin */, "foo3").then((result) => {
        return new Promise((resolve) => {
            window.onmessage = () => {
                window.onmessage = null;
                resolve();
            }

            let destinationURL = get_host_info().HTTP_ORIGIN + "/WebKit/cross-origin-options/resources/destination.html";
            w = open(destinationURL, "foo3");
            assert_equals(w, result.window, "open() should return the same window");
        });
    });
}, "'Cross-Origin-Options: allow' does not prevent navigation from opener via open() target");
</script>
</body>
</html>
